Implement new results api

+++ driver/src/main/java/org/neo4j/driver/v1/Path.java
@@ -61,7 +61,7 @@ public interface Path extends Iterable<Path.Segment>
     Node end();
 
     /** @return the number of segments in this path, which will be the same as the number of relationships */
-    long length();
+    int length();
 
     /**
      * @param node the node to check for
